ObjectSpawner Update
Rename BaseSpawnable to Spawnable.
Add YOffset support for Spawnable types. This calculates your collider offset for spawning.
Add ISpawnable.
Add BaseRespawnCondition.
Add DeadNPCRespawnCondition.
Interactable ID is now an int64.
Update SceneObjectNamer, added server authority over the names.
Rename SceneObjectUID to SceneObject.
Add ISceneObject.
SceneObject now allows objects to register and obtain a UID.
Update IntBitExtensions and LongBitExtensions, they should no longer box.
Add WorldItem.
ObjectSpawner is now functional.
Fixed an issue with Factions not calculating properly.

+++ FishMMO-Unity/Assets/Scripts/Client/UI/Controls/World/Ability/Crafting/UIAbilityCraft.cs
@@ -17,7 +17,7 @@ namespace FishMMO.Client
 		public UITooltipButton AbilityEventPrefab;
 		public CharacterAttributeTemplate CurrencyTemplate;
 
-		private int lastInteractableID = 0;
+		private long lastInteractableID = 0;
 		private Dictionary<int, UITooltipButton> EventSlots;
 
 		public override void OnStarting()
