+++ FishMMO-Unity/Assets/Scripts/Shared/Entity/Item/Container/Bank/IBankController.cs
@@ -2,7 +2,7 @@
 {
 	public interface IBankController : ICharacterBehaviour, IItemContainer
 	{
-		int LastInteractableID { get; set; }
+		long LastInteractableID { get; set; }
 		long Currency { get; set; }
 		bool CanSwapItemSlots(int from, int to, InventoryType fromInventory);
 	}
